Delete API method and API alias metthod if exists

+++ scripts/14_create_lambda_api_method.sh
@@ -6,9 +6,7 @@ aws s3 cp ~/${LAMBDA_ZIP} s3://${S3_BUCKET}/lambda/
 
 PRE_EX_LAMBDA=$(aws lambda get-function-configuration \
                                 --function-name  ${LAMBDA_FUNCTION_NAME} \
-                                --query FunctionName)
-echo -e "${PRE_EX_LAMBDA}"    
-echo -e "\"${LAMBDA_FUNCTION_NAME}\""           
+                                --query FunctionName)      
                     
 # Delete lambda function if already exists
  
@@ -52,13 +50,21 @@ echo -e "$INFO API_ARN: $(FY $API_ARN)"
 echo
 echo -e "$INFO API method for Resource $(FC $API_RESOURCE_NAME)"
 
+API_METHOD=$(aws apigateway get-method --rest-api-id ${API_ID} \
+               --resource-id ${API_RESOURCE_ID} \
+               --http-method ${API_HTTP_METHOD} \
+                --query httpMethod)
+                
 # Delete API method if already exists
-echo -e "INFO  Deleting pre-existent HTTP method"
-aws apigateway delete-method \
-    --rest-api-id ${API_ID} \
-    --resource-id ${API_RESOURCE_ID} \
-    --http-method ${API_HTTP_METHOD} \
-    --output table
+if [[ "${API_METHOD}" ==  "\"${API_HTTP_METHOD}\"" ]]; then
+       echo -e "$INFO  Deleting pre-existent HTTP method"
+       aws apigateway delete-method \
+            --rest-api-id ${API_ID} \
+            --resource-id ${API_RESOURCE_ID} \
+            --http-method ${API_HTTP_METHOD} \
+            --output table 
+fi
+
 
 
 # Creating API method under specified resource
@@ -127,14 +133,20 @@ echo -e "$INFO Finished creating $(FC $API_HTTP_METHOD) under" \
 echo
 echo -e "$INFO API method for Resource $(FC $API_ALIAS_RESOURCE_NAME)"
 
+API_ALIAS_METHOD=$(aws apigateway get-method --rest-api-id ${API_ID} \
+               --resource-id ${API_ALIAS_RESOURCE_ID} \
+               --http-method ${API_HTTP_METHOD} \
+                --query httpMethod)
+                
 # Delete API method if already exists
-echo -e "INFO  Deleting pre-existent HTTP method"
-aws apigateway delete-method \
-    --rest-api-id ${API_ID} \
-    --resource-id ${API_ALIAS_RESOURCE_ID} \
-    --http-method ${API_HTTP_METHOD} \
-    --output table
-
+if [[ "${API_ALIAS_METHOD}" ==  "\"${API_HTTP_METHOD}\"" ]]; then
+       echo -e "$INFO  Deleting pre-existent HTTP method"
+       aws apigateway delete-method \
+            --rest-api-id ${API_ID} \
+            --resource-id ${API_ALIAS_RESOURCE_ID} \
+            --http-method ${API_HTTP_METHOD} \
+            --output table 
+fi
 
 # Creating API method under specified resource
 echo -e "$INFO Creating ${API_HTTP_METHOD} under ${API_ALIAS_RESOURCE_NAME} resource"
